YARN-2607. Fixed issues in TestDistributedShell. Contributed by Wangda Tan.
(cherry picked from commit 737d9284c109dac20ff423f30c62f6abe2db10f7)
